+++ docs/changelog.rst
@@ -5,6 +5,9 @@ Changelog
 ------------------
 
 - #94 Modify Signatures with datetime for Tested by and Verified by in final report
+- #87 Remove preliminary comment
+- #85 Add Result Variable's unit in final report
+- #75 Make Department field required based on Client in Sample View
 - #82 Remove "Results" heading from final report
 - #79 Update consumer when call pull tamanu patients
 - #60 Add analysts name for report comments